+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `TextInputWithUnit`: Add `secondUnit` prop.
 - Feature: `HorizontalCrowdfundingCard`: Add `noProgress` prop, defaults to `false`.
 
 ## [7.0.0] - 2022-01-18